script without rsa key
This script takes a fourth argument, SSH_PASSWORD, which is the password for the SSH connection. Instead of using the -i option to specify the private key, it uses sshpass to pass the password as an argument to the ssh and scp commands. The -o StrictHostKeyChecking=no option is still used to disable strict host key checking.